refactor: break handler classes into separate files
- Extract TextContentHandler into TextContentHandler.ts
- Extract ParameterHandler into ParameterHandler.ts
- Extract ToolUseHandler into ToolUseHandler.ts
- Create types.ts for shared type definitions
- Update StreamingParser to import from separate files
- Update index.ts exports to reference correct files
- Update README.md to reflect new file structure
- Maintain backward compatibility and streaming behavior
- All tests pass, multi-contributor friendly structure
