+++ src/core/assistant-message/directives/ParameterHandler.ts
@@ -1,0 +1,22 @@
+import { ParsingState } from "./types"
+
+export class ParameterHandler {
+	static handleParameter(state: ParsingState): boolean {
+		if (!state.currentToolUse || !state.currentParamName) return false
+
+		const currentParamValue = state.accumulator.slice(state.currentParamValueStartIndex)
+		const paramClosingTag = `</${state.currentParamName}>`
+
+		if (currentParamValue.endsWith(paramClosingTag)) {
+			// End of param value.
+			state.currentToolUse.params[state.currentParamName] = currentParamValue
+				.slice(0, -paramClosingTag.length)
+				.trim()
+			state.currentParamName = undefined
+			return true
+		} else {
+			// Partial param value is accumulating.
+			return true
+		}
+	}
+}
