+++ src/core/assistant-message/directives/index.ts
@@ -1,2 +1,5 @@
-export { StreamingParser, TextContentHandler, ToolUseHandler, ParameterHandler } from "./StreamingParser"
-export type { TextDirective, ToolDirective, Directive, ParsingState } from "./StreamingParser"
+export { StreamingParser } from "./StreamingParser"
+export { TextContentHandler } from "./TextContentHandler"
+export { ToolUseHandler } from "./ToolUseHandler"
+export { ParameterHandler } from "./ParameterHandler"
+export type { TextDirective, ToolDirective, Directive, ParsingState } from "./types"
